# openapi.yaml
openapi: 3.1.0
info:
  title: DriveMeMaybeAPI
  version: 0.1.0
  description: |
    REST API for DriveMeMaybe, a mobile app for tracking shared trips and keeping
    driving balances fair. Manages users, trip groups (invite-by-code), group members,
    trips (driver + passengers, distance and/or duration), per-member balances and
    group driving-time insights. Consumed primarily by the DriveMeMaybe mobile app,
    usable by any HTTP client.
  license:
    name: GPL-3.0-only
    url: https://www.gnu.org/licenses/gpl-3.0.html
servers:
  - url: http://localhost:8080
    description: Development (default Spring Boot port)
  - url: https://api.example.com
    description: Production (replace with real host)
...
paths:
  /auth/signup:
    post:
      tags: [Authentication]
      summary: Register a new user
      operationId: signupUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDTO'
            examples:
              default:
                value:
                  email: alice@example.com
                  password: secret123
                  name: Alice
                  nickname: ali
                  pfp: https://example.com/pfp/alice.png
                  preferences:
                    isDarkMode: false
                    preferredLang: EN
      responses:
        '200':
          description: Registered. Returns JWT and expiry.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
              examples:
                default:
                  value:
                    token: eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJhbGljZUBleGFtcGxlLmNvbSJ9.signature
                    expiresOn: 1789990800000
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /auth/login:
    post:
      tags: [Authentication]
      summary: Login with email and password
      operationId: loginUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLoginDTO'
            examples:
              default:
                value:
                  email: alice@example.com
                  password: secret123
      responses:
        '200':
          description: Authenticated. Returns JWT and expiry.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
              examples:
                default:
                  value:
                    token: eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJhbGljZUBleGFtcGxlLmNvbSJ9.signature
                    expiresOn: 1789990800000
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /auth/refresh:
    get:
      tags: [Authentication]
      summary: Re-issue a token for the current user
      operationId: refreshToken
      responses:
        '200':
          description: New token issued.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
              examples:
                default:
                  value:
                    token: eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJhbGljZUBleGFtcGxlLmNvbSJ9.signature
                    expiresOn: 1789990800000
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /auth/google/login:
    post:
      tags: [Authentication]
      summary: Login with a Google ID token
      operationId: loginGoogleUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GoogleLoginDTO'
            examples:
              default:
                value:
                  idToken: eyJhbGciOiJSUzI1NiIsImtpZCI6IjEyMzQifQ.eyJlbWFpbCI6ImFsaWNlQGV4YW1wbGUuY29tIn0.signature
      responses:
        '200':
          description: Authenticated via Google. Returns JWT and expiry.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
              examples:
                default:
                  value:
                    token: eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJhbGljZUBleGFtcGxlLmNvbSJ9.signature
                    expiresOn: 1789990800000
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /groups/group:
    post:
      tags: [Groups]
      summary: Create a trip group
      operationId: createGroup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GroupRequestDTO'
            examples:
              default:
                value:
                  name: Weekend road crew
                  pfp: https://example.com/pfp/group.png
      responses:
        '200':
          description: Created group (API returns 200, not 201).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupDTO'
              examples:
                default:
                  value:
                    id: aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaaa
                    name: Weekend road crew
                    groupCode: XK7Q2M
                    groupMembers: 1
                    groupTrips: 0
                    pfp: https://example.com/pfp/group.png
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /groups/{groupId}:
    parameters:
      - $ref: '#/components/parameters/GroupIdPath'
    get:
      tags: [Groups]
      summary: Get group details (members only)
      operationId: getGroupDetail
      responses:
        '200':
          description: The group.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupDTO'
              examples:
                default:
                  value:
                    id: aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaaa
                    name: Weekend road crew
                    groupCode: XK7Q2M
                    groupMembers: 2
                    groupTrips: 5
                    pfp: https://example.com/pfp/group.png
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /groups/{groupId}/members:
    parameters:
      - $ref: '#/components/parameters/GroupIdPath'
    get:
      tags: [Groups]
      summary: List group members
      operationId: listGroupMembers
      responses:
        '200':
          description: Member list.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GroupMemberDTO'
              examples:
                default:
                  value:
                    - id: 11111111-1111-1111-1111-111111111111
                      nickname: ali
                    - id: 22222222-2222-2222-2222-222222222222
                      nickname: bob
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /groups/{groupId}/members/balance:
    parameters:
      - $ref: '#/components/parameters/GroupIdPath'
    get:
      tags: [Groups]
      summary: Get per-member driving balances
      operationId: listGroupMemberBalances
      responses:
        '200':
          description: Balances. Driving increases balance, riding decreases it.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GroupMemberBalanceDTO'
              examples:
                default:
                  value:
                    - nickname: ali
                      timeBalance: 45
                      kmBalance: 50
                    - nickname: bob
                      timeBalance: -45
                      kmBalance: -50
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /groups/{groupId}/trips:
    parameters:
      - $ref: '#/components/parameters/GroupIdPath'
    get:
      tags: [Groups]
      summary: List trips in a group
      operationId: listGroupTrips
      responses:
        '200':
          description: Trips in the group.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TripRequestDTO'
              examples:
                default:
                  value:
                    - driver: ali
                      date: '2026-09-17'
                      durationMinutes: 45
                      origin: Berlin
                      destination: Potsdam
                      notes: Evening return
                      tripId: cccccccc-cccc-cccc-cccc-cccccccccccc
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /groups/{groupId}/insights:
    parameters:
      - $ref: '#/components/parameters/GroupIdPath'
    post:
      tags: [Groups]
      summary: Get driving-time insights for a period
      operationId: getGroupInsights
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GroupInsightsRequest'
            examples:
              monthly:
                value:
                  periodType: MONTHLY
                  year: 2026
                  month: 9
              yearly:
                value:
                  periodType: YEARLY
                  year: 2026
                  month: null
      responses:
        '200':
          description: Aggregated driving time per driver nickname.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupInsights'
              examples:
                default:
                  value:
                    year: 2026
                    month: 9
                    driverDrivingTime:
                      ali: 120
                      bob: 45
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /groups/{groupId}/update:
    parameters:
      - $ref: '#/components/parameters/GroupIdPath'
    put:
      tags: [Groups]
      summary: Update a group
      operationId: updateGroup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GroupDTO'
            examples:
              default:
                value:
                  id: aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaaa
                  name: Weekend road crew 2026
                  groupCode: XK7Q2M
                  groupMembers: 2
                  groupTrips: 5
                  pfp: https://example.com/pfp/group-new.png
      responses:
        '200':
          description: Updated group.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupDTO'
              examples:
                default:
                  value:
                    id: aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaaa
                    name: Weekend road crew 2026
                    groupCode: XK7Q2M
                    groupMembers: 2
                    groupTrips: 5
                    pfp: https://example.com/pfp/group-new.png
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /groups/{groupId}/leave:
    parameters:
      - $ref: '#/components/parameters/GroupIdPath'
    put:
      tags: [Groups]
      summary: Leave a group
      operationId: leaveGroup
      responses:
        '200':
          description: Left the group.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Send `Authorization: Bearer <token>`. Tokens are issued by
        POST /auth/signup, POST /auth/login and POST /auth/google/login
        (1 hour lifetime) and can be re-issued via GET /auth/refresh.
  parameters:
    GroupIdPath:
      name: groupId
      in: path
      required: true
      description: Group UUID.
      schema:
        type: string
        format: uuid
      example: aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaaa
...
  schemas:
    UserDTO:
      type: object
      description: Signup / profile-update payload. `password` and `pfp` are optional. `preferences.preferredLang` is resolved case-insensitively against the languages table (defaults to `EN`); `preferences.isDarkMode` maps to the stored dark-mode flag.
      required: [email, name, nickname]
      properties:
        email:
          type: string
          format: email
          example: alice@example.com
        password:
          type: string
          format: password
          minLength: 8
          example: secret123
        name:
          type: string
          minLength: 1
          example: Alice
        nickname:
          type: string
          minLength: 1
          example: ali
        pfp:
          type: string
          format: uri
          example: https://example.com/pfp/alice.png
        preferences:
          $ref: '#/components/schemas/UserPreferences'
    UserLoginDTO:
      type: object
      description: Email/password login payload.
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: alice@example.com
        password:
          type: string
          format: password
          example: secret123
...
    UserPreferences:
      type: object
      description: User display preferences. `isDarkMode` maps to `PREF_DARK_MODE`; `preferredLang` carries the language code on write and the formal name on read.
      properties:
        isDarkMode:
          type: boolean
          nullable: true
          example: true
        preferredLang:
          type: string
          nullable: true
          example: EN
...
    GoogleLoginDTO:
      type: object
      description: Google Sign-In payload; `idToken` is verified server-side.
      required: [idToken]
      properties:
        idToken:
          type: string
          example: eyJhbGciOiJSUzI1NiIsImtpZCI6IjEyMzQifQ.eyJlbWFpbCI6ImFsaWNlQGV4YW1wbGUuY29tIn0.signature
    LoginResponse:
      type: object
      description: JWT plus absolute expiry in epoch millis.
      required: [token, expiresOn]
      properties:
        token:
          type: string
          example: eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJhbGljZUBleGFtcGxlLmNvbSJ9.signature
        expiresOn:
          type: integer
          format: int64
          description: Epoch milliseconds when the token expires (issued + 3600000 ms).
          example: 1789990800000
    GroupDTO:
      type: object
      description: Trip group. `pfp` is optional. `groupMembers` / `groupTrips` are derived counts (members.size() / trips.size()).
      required: [id, name, groupCode]
      properties:
        id:
          type: string
          format: uuid
          example: aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaaa
        name:
          type: string
          minLength: 1
          example: Weekend road crew
        groupCode:
          type: string
          example: XK7Q2M
        groupMembers:
          type: integer
          minimum: 0
          example: 2
        groupTrips:
          type: integer
          minimum: 0
          example: 5
        pfp:
          type: string
          format: uri
          nullable: true
          example: https://example.com/pfp/group.png
    GroupRequestDTO:
      type: object
      description: Create-group payload.
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          example: Weekend road crew
        pfp:
          type: string
          format: uri
          example: https://example.com/pfp/group.png
    GroupMemberDTO:
      type: object
      description: Group membership entry.
      required: [id, nickname]
      properties:
        id:
          type: string
          format: uuid
          example: 11111111-1111-1111-1111-111111111111
        nickname:
          type: string
          example: ali
    GroupMemberBalanceDTO:
      type: object
      description: Accumulated balance per member. Driving adds, riding subtracts.
      required: [nickname, timeBalance, kmBalance]
      properties:
        nickname:
          type: string
          example: ali
        timeBalance:
          type: integer
          description: Net minutes.
          example: 45
        kmBalance:
          type: integer
          description: Net kilometers.
          example: 50
    GroupInsightsRequest:
      type: object
      description: Period selector. `month` is required for MONTHLY, null/omitted for YEARLY.
      required: [periodType, year]
      properties:
        periodType:
          type: string
          enum: [MONTHLY, YEARLY]
          example: MONTHLY
        year:
          type: integer
          minimum: 2000
          maximum: 2100
          example: 2026
        month:
          type: integer
          minimum: 1
          maximum: 12
          nullable: true
          example: 9
    GroupInsights:
      type: object
      description: Driving-time aggregates per driver nickname for the period.
      required: [year, driverDrivingTime]
      properties:
        year:
          type: integer
          example: 2026
        month:
          type: integer
          nullable: true
          minimum: 1
          maximum: 12
          example: 9
        driverDrivingTime:
          type: object
          description: Map of driver nickname to driven minutes.
          additionalProperties:
            type: integer
            minimum: 0
          example:
            ali: 120
            bob: 45
...
    TripRequestDTO:
      type: object
      description: Trip list entry returned by GET /groups/{groupId}/trips.
      required: [driver, date, tripId]
      properties:
        driver:
          type: string
          description: Driver nickname.
          example: ali
        date:
          type: string
          format: date
          example: '2026-09-17'
        durationMinutes:
          type: integer
          nullable: true
          example: 45
        origin:
          type: string
          nullable: true
          example: Berlin
        destination:
          type: string
          nullable: true
          example: Potsdam
        notes:
          type: string
          nullable: true
          example: Evening return
        tripId:
          type: string
          format: uuid
          example: cccccccc-cccc-cccc-cccc-cccccccccccc
    Error:
      type: object
      description: |
        Central error envelope from CustomExceptionHandler (CustomErrorResponse).
        NOTE: the current handler always constructs the body with NOT_FOUND
        (statusCode 404 / statusName NOT_FOUND) while the HTTP status line
        carries the real code — examples below show the intended contract.
      required: [errorCode, detail, statusCode, statusName, timestamp]
      properties:
        errorCode:
          type: string
          description: Machine-readable code, e.g. validation-exception, authorizationException.
          example: validation-exception
        detail:
          type: string
          example: Group not found
        statusCode:
          type: integer
          example: 404
        statusName:
          type: string
          example: NOT_FOUND
        timestamp:
          type: string
          format: date-time
          example: '2026-09-17T12:00:00'
  responses:
    BadRequest:
      description: Validation / malformed request (ValidationException → HTTP 400).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            default:
              value:
                errorCode: validation-exception
                detail: Group name must not be blank
                statusCode: 400
                statusName: BAD_REQUEST
                timestamp: '2026-09-17T12:00:00'
    Unauthorized:
      description: Authentication failed (AuthenticationException → HTTP 401).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            default:
              value:
                errorCode: authentication-exception
                detail: Invalid email or password
                statusCode: 401
                statusName: UNAUTHORIZED
                timestamp: '2026-09-17T12:00:00'
    Forbidden:
      description: Missing, invalid or expired JWT (JwtException / BadCredentialsException → HTTP 403 with errorCode authorizationException).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            default:
              value:
                errorCode: authorizationException
                detail: JWT expired
                statusCode: 403
                statusName: FORBIDDEN
                timestamp: '2026-09-17T12:00:00'
    NotFound:
      description: Entity missing (NoEntitiesFoundException → HTTP 404).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            default:
              value:
                errorCode: no-entities-found
                detail: Group aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaaa not found
                statusCode: 404
                statusName: NOT_FOUND
                timestamp: '2026-09-17T12:00:00'
    InternalError:
      description: Server failure (DatabaseException / UnexpectedException / generic fallback → HTTP 500).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            default:
              value:
                errorCode: generic-exception
                detail: An unexpected error occurred, we are working on it.
                statusCode: 500
                statusName: INTERNAL_SERVER_ERROR
                timestamp: '2026-09-17T12:00:00'
security:
  - bearerAuth: []